check if history file exists

+++ System/Cli.php
@@ -99,10 +99,16 @@ class Cli
      */
     public static function init(): void
     {
-        file_put_contents(self::$history, json_encode([
-            'date' => (new \DateTime())->format('Y-m-d\TH:i:s\Z'),
-            'changes' => [],
-        ], JSON_PRETTY_PRINT|JSON_UNESCAPED_SLASHES));
+        if (!file_exists(self::$history)) {
+            file_put_contents(self::$history, json_encode([
+                'date' => (new \DateTime())->format('Y-m-d\TH:i:s\Z'),
+                'changes' => [],
+            ], JSON_PRETTY_PRINT|JSON_UNESCAPED_SLASHES));
+
+            die('[+] History file has been created!');
+        }
+
+        die('[!] History file already exists!');
     }
 
     /**
